requirements.txt
langchain-groq
python-dotenv
langchain_community
langchain-elasticsearch
sentence-transformers
langchain-huggingface